ai-backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
requests
langgraph
langchain-google-genai
langchain-core
langchain-community
